refactor: Melhoria no retorno de erros
- Agora os erros dos DTOS retornam as mensagens em português, o que facilita a captura de erros no front
