+++ api/src/main/java/com/dbserver/votingchallenge/dtos/agenda/AgendaOpenVotingSessionDTO.java
@@ -1,0 +1,11 @@
+package com.dbserver.votingchallenge.dtos.agenda;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Builder;
+
+@Builder
+@Schema(name = "Abrir uma sessão de votação")
+public record AgendaOpenVotingSessionDTO(
+        Integer timeInMinutes
+) {
+}
