+++ api/src/main/java/com/dbserver/votingchallenge/dtos/agenda/AgendaResponseListDTO.java
@@ -1,5 +1,7 @@
 package com.dbserver.votingchallenge.dtos.agenda;
 
+import com.dbserver.votingchallenge.dtos.general.StatusDTO;
+import com.dbserver.votingchallenge.dtos.votingSession.VotingSessionResultDTO;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
@@ -8,6 +10,8 @@ import lombok.Builder;
 public record AgendaResponseListDTO(
         Integer id,
         String name,
-        String description
+        String description,
+        StatusDTO status,
+        VotingSessionResultDTO results
 ) {
 }
