+++ api/src/main/java/com/dbserver/votingchallenge/dtos/agenda/AgendaVoteDTO.java
@@ -1,15 +1,16 @@
-package com.dbserver.votingchallenge.dtos.votingSession;
+package com.dbserver.votingchallenge.dtos.agenda;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
 @Builder
 @Schema(name = "Votar em uma pauta")
-public record VotingSessionVoteDTO(
-        @NotNull
+public record AgendaVoteDTO(
+        @NotNull(message = "É obrigatório")
         Boolean favorable,
-        @NotNull
+        @NotBlank(message = "Não pode ficar em branco")
         String associatedCpf
 ) {
 }
