+++ api/src/main/java/com/dbserver/votingchallenge/dtos/general/ErrorResponseDTO.java
@@ -1,4 +1,4 @@
 package com.dbserver.votingchallenge.dtos.general;
 
-public record ErrorResponseDTO(String error) {
+public record ErrorResponseDTO(String baseError) {
 }
