+++ api/src/main/java/com/dbserver/votingchallenge/dtos/general/RequestBodyErrorResponseDTO.java
@@ -5,10 +5,10 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import java.util.List;
 
 public class RequestBodyErrorResponseDTO {
-    public List<RequestBodyErrorDetailDTO> errors;
+    public List<RequestBodyErrorDetailDTO> fieldErrors;
 
     public RequestBodyErrorResponseDTO(MethodArgumentNotValidException exception) {
-        this.errors = exception.getFieldErrors().stream().map(error ->
+        this.fieldErrors = exception.getFieldErrors().stream().map(error ->
                 new RequestBodyErrorDetailDTO(error.getField(), error.getDefaultMessage())
         ).toList();
     }
