+++ api/src/main/java/com/dbserver/votingchallenge/dtos/vote/VoteCreateDTO.java
@@ -1,14 +1,15 @@
 package com.dbserver.votingchallenge.dtos.vote;
 
 import com.dbserver.votingchallenge.domain.associated.Associated;
+import com.dbserver.votingchallenge.domain.votingSession.VotingSession;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
 @Builder
-@Schema(name = "Retorno de um voto")
-public record VoteResponseDTO(
-        Integer id,
+@Schema(name = "Criar um voto")
+public record VoteCreateDTO(
         Boolean favorable,
-        Associated associated
+        Associated associated,
+        VotingSession votingSession
 ) {
 }
